PHPStan (Level 0) on classes/

+++ classes/module/Module.php
@@ -417,7 +417,9 @@ abstract class ModuleCore implements ModuleInterface
         $result = Db::getInstance()->insert($this->table, ['name' => $this->name, 'active' => 1, 'version' => $this->version]);
         if (!$result) {
             $this->_errors[] = Context::getContext()->getTranslator()->trans('Technical error: PrestaShop could not install this module.', [], 'Admin.Modules.Notification');
-            $this->uninstallTabs();
+            if (method_exists($this, 'uninstallTabs')) {
+                $this->uninstallTabs();
+            }
             $this->uninstallOverrides();
 
             return false;
@@ -1273,7 +1275,9 @@ abstract class ModuleCore implements ModuleInterface
         global $_MODULES;
         $file = _PS_MODULE_DIR_ . $module . '/' . Context::getContext()->language->iso_code . '.php';
         if (Tools::file_exists_cache($file) && include_once($file)) {
+            /* @phpstan-ignore-next-line Defined variable in translation file */
             if (isset($_MODULE) && is_array($_MODULE)) {
+                /** @phpstan-ignore-next-line Defined variable in translation file */
                 $_MODULES = !empty($_MODULES) ? array_merge($_MODULES, $_MODULE) : $_MODULE;
             }
         }
@@ -1372,7 +1376,9 @@ abstract class ModuleCore implements ModuleInterface
                 if (!count($module_errors) && (int) $xml_module->need_instance == 0) {
                     $file = _PS_MODULE_DIR_ . $module . '/' . Context::getContext()->language->iso_code . '.php';
                     if (Tools::file_exists_cache($file) && include_once($file)) {
+                        /* @phpstan-ignore-next-line Defined variable in translation file */
                         if (isset($_MODULE) && is_array($_MODULE)) {
+                            /** @phpstan-ignore-next-line Defined variable in translation file */
                             $_MODULES = !empty($_MODULES) ? array_merge($_MODULES, $_MODULE) : $_MODULE;
                         }
                     }
@@ -2140,8 +2146,6 @@ abstract class ModuleCore implements ModuleInterface
 
         // Close div openned previously
         $output .= '</div></div>';
-
-        $this->error = true;
 
         return $output;
     }
@@ -2912,15 +2916,6 @@ abstract class ModuleCore implements ModuleInterface
      */
     public function getPosition($id_hook)
     {
-        if (isset(Hook::$preloadModulesFromHooks)) {
-            if (isset(Hook::$preloadModulesFromHooks[$id_hook])) {
-                if (isset(Hook::$preloadModulesFromHooks[$id_hook]['module_position'][$this->id])) {
-                    return Hook::$preloadModulesFromHooks[$id_hook]['module_position'][$this->id];
-                } else {
-                    return 0;
-                }
-            }
-        }
         $result = Db::getInstance()->getRow('
             SELECT `position`
             FROM `' . _DB_PREFIX_ . 'hook_module`
